Move shared constants interfaces to Shared folder
