+++ src/php/Compiler/Domain/Analyzer/Environment/GlobalEnvironment.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Phel\Compiler\Domain\Analyzer\Environment;
 
-use Phel\Build\BuildConstants;
 use Phel\Compiler\Domain\Analyzer\Ast\AbstractNode;
 use Phel\Compiler\Domain\Analyzer\Ast\GlobalVarNode;
 use Phel\Compiler\Domain\Analyzer\Ast\LiteralNode;
@@ -15,6 +14,8 @@ use Phel\Lang\Registry;
 use Phel\Lang\SourceLocation;
 use Phel\Lang\Symbol;
 use Phel\Lang\TypeFactory;
+use Phel\Shared\BuildConstants;
+use Phel\Shared\CompilerConstants;
 use RuntimeException;
 
 use function array_key_exists;
@@ -22,8 +23,6 @@ use function dirname;
 
 final class GlobalEnvironment implements GlobalEnvironmentInterface
 {
-    private const PHEL_CORE_NAMESPACE = 'phel\core';
-
     private string $ns = 'user';
 
     /** @var array<string, array<string, bool>> */
@@ -45,7 +44,7 @@ final class GlobalEnvironment implements GlobalEnvironmentInterface
 
     public function __construct()
     {
-        $this->addInternalCompileModeDefinition();
+        $this->addInternalBuildModeDefinition();
     }
 
     public function getNs(): string
@@ -160,7 +159,6 @@ final class GlobalEnvironment implements GlobalEnvironmentInterface
         }
 
         if (isset($this->useAliases[$this->ns][$strName])) {
-            /** @var Symbol $alias */
             $alias = $this->useAliases[$this->ns][$strName];
             $alias->copyLocationFrom($name);
             return new PhpClassNameNode($env, $alias, $name->getStartLocation());
@@ -194,15 +192,15 @@ final class GlobalEnvironment implements GlobalEnvironmentInterface
         $this->interfaces[$namespace][$name->getName()] = $name;
     }
 
-    private function addInternalCompileModeDefinition(): void
+    private function addInternalBuildModeDefinition(): void
     {
         $symbol = Symbol::create(BuildConstants::BUILD_MODE);
         $meta = TypeFactory::getInstance()->persistentMapFromKVs(
             Keyword::create('doc'),
             'Set to true when a file is compiled, false otherwise.',
         );
-        Registry::getInstance()->addDefinition(self::PHEL_CORE_NAMESPACE, $symbol->getName(), false, $meta);
-        $this->addDefinition(self::PHEL_CORE_NAMESPACE, $symbol);
+        Registry::getInstance()->addDefinition(CompilerConstants::PHEL_CORE_NAMESPACE, $symbol->getName(), false, $meta);
+        $this->addDefinition(CompilerConstants::PHEL_CORE_NAMESPACE, $symbol);
     }
 
     private function resolveMagicFile(?SourceLocation $sl): ?string
@@ -253,7 +251,7 @@ final class GlobalEnvironment implements GlobalEnvironmentInterface
         }
 
         return $this->resolveInterfaceOrDefinitionForCurrentNs($name, $env, $currentNs)
-            ?? $this->resolveInterfaceOrDefinition($name, $env, self::PHEL_CORE_NAMESPACE);
+            ?? $this->resolveInterfaceOrDefinition($name, $env, CompilerConstants::PHEL_CORE_NAMESPACE);
     }
 
     /**
