+++ src/php/Shared/CompilerConstants.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phel\Shared;
+
+interface CompilerConstants
+{
+    public const PHEL_CORE_NAMESPACE = 'phel\core';
+}
